Handle message action edge cases
