+++ mobile/lib/features/channels/message_actions.dart
@@ -39,6 +39,7 @@ import 'thread_follows/thread_follows_provider.dart';
 import 'timeline_message.dart';
 
 part 'message_actions/reaction_popover.dart';
+part 'message_actions/quick_reaction_row.dart';
 part 'message_actions/message_action_popover.dart';
 
 /// Preview length for reminder targets — matches desktop's
@@ -663,127 +664,9 @@ class _FastActionTile extends StatelessWidget {
   }
 }
 
-/// The row of one-tap reactions at the top of the action sheet, plus the "+"
-/// tile that opens the full picker.
-///
 /// The emoji shown are the user's own frequently-used set (desktop's
 /// `useQuickReactionEmojis` behaviour), topped up with [defaultQuickEmojis] so
 /// the row is full on a fresh install.
-class _QuickReactionRow extends ConsumerWidget {
-  final TimelineMessage message;
-
-  /// The sheet's context, popped before the reaction fires.
-  final BuildContext sheetContext;
-
-  /// The long-pressed message's page context — survives the sheet pop, so the
-  /// picker opened from "+" isn't torn down with the sheet.
-  final BuildContext pageContext;
-
-  /// The long-pressed message's page ref. The picker callback outlives this
-  /// bottom sheet, so it must not read through the sheet's disposed ref.
-  final WidgetRef pageRef;
-
-  /// Drives the staged glyph reveal when this row is shown in the popover.
-  /// The bottom sheet leaves this null and retains its existing static row.
-  final Animation<double>? presentationAnimation;
-
-  const _QuickReactionRow({
-    required this.message,
-    required this.sheetContext,
-    required this.pageContext,
-    required this.pageRef,
-    this.presentationAnimation,
-  });
-
-  @override
-  Widget build(BuildContext context, WidgetRef ref) {
-    final customEmoji = ref.watch(customEmojiListProvider);
-    final emoji = quickReactionEmoji(
-      ref.watch(recentEmojiProvider),
-      customShortcodes: {
-        for (final entry in customEmoji) entry.shortcode.toLowerCase(),
-      },
-    );
-    final customByShortcode = {
-      for (final entry in customEmoji) entry.shortcode.toLowerCase(): entry,
-    };
-
-    void react(String value) {
-      // The generic picker is also used for composing and statuses. Record
-      // recency here, at the reaction call site, so only reactions drive the
-      // quick-reaction row.
-      pageRef.read(recentEmojiProvider.notifier).record(value);
-      // The sheet is on its way out, so the burst can't come from this tile —
-      // hand it to the pill that's about to appear in the timeline.
-      armReactionBurst(pageRef, message, value);
-      pageRef.read(channelActionsProvider).addReaction(message.id, value);
-    }
-
-    return LayoutBuilder(
-      builder: (context, constraints) {
-        const desiredCircleSize = 52.0;
-        const minimumCircleSize = 44.0;
-        final itemCount = emoji.length + 1;
-        final gapCount = itemCount - 1;
-        final circleSize =
-            ((constraints.maxWidth - (Grid.twelve * gapCount)) / itemCount)
-                .clamp(minimumCircleSize, desiredCircleSize)
-                .toDouble();
-        final gap =
-            ((constraints.maxWidth - (circleSize * itemCount)) / gapCount)
-                .clamp(0.0, Grid.twelve)
-                .toDouble();
-        final circles = <Widget>[
-          for (var index = 0; index < emoji.length; index++)
-            _ReactionItemReveal(
-              key: ValueKey('quick-reaction-${emoji[index]}'),
-              animation: presentationAnimation,
-              index: index,
-              child: _QuickReactionCircle(
-                size: circleSize,
-                onTap: () {
-                  Navigator.of(sheetContext).pop();
-                  react(emoji[index]);
-                },
-                child: _QuickReactionGlyph(
-                  value: emoji[index],
-                  customByShortcode: customByShortcode,
-                ),
-              ),
-            ),
-          _ReactionItemReveal(
-            key: const ValueKey('quick-reaction-more'),
-            animation: presentationAnimation,
-            index: emoji.length,
-            child: _QuickReactionCircle(
-              size: circleSize,
-              onTap: () {
-                Navigator.of(sheetContext).pop();
-                showEmojiPicker(context: pageContext, onSelect: react);
-              },
-              child: Icon(
-                LucideIcons.plus,
-                size: 24,
-                color: context.colors.onSurfaceVariant,
-              ),
-            ),
-          ),
-        ];
-
-        return Row(
-          mainAxisAlignment: MainAxisAlignment.center,
-          children: [
-            for (var index = 0; index < circles.length; index++) ...[
-              circles[index],
-              if (index < circles.length - 1) SizedBox(width: gap),
-            ],
-          ],
-        );
-      },
-    );
-  }
-}
-
 class _ReactionItemReveal extends StatelessWidget {
   final Animation<double>? animation;
   final int index;
